requirements.txt
pytest
google-genai
pydantic
python-dotenv